requirements.txt
flask
groq
markdown2
python-dotenv